Multilingual. However, I have to use IPageTexts.d.ts instead of IPageTexts.ts. Otherwise, Angular will show a compiling error.

+++ src/app/services/fs.service.ts
@@ -138,6 +138,11 @@ export class FsService {
 
   rmFile$(file: FileEntry) {
     return new Observable<boolean>( subs => {
+      if (!!file === false) {
+        subs.next(false);
+        subs.complete();
+        return;
+      }
       file.remove(
         () => {
           subs.next(true);
